Move errors into design module
