+++ design/src/main/kotlin/uk/govuk/app/design/ui/component/error/ErrorConstants.kt
@@ -1,0 +1,5 @@
+package uk.govuk.app.design.ui.component.error
+
+object ErrorConstants {
+    const val GOV_UK_URL = "https://www.gov.uk"
+}
