+++ design/src/main/kotlin/uk/govuk/app/design/ui/component/error/OfflineMessage.kt
@@ -1,13 +1,12 @@
-package uk.govuk.app.networking.ui.component
+package uk.govuk.app.design.ui.component.error
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import uk.govuk.app.design.ui.component.Error
 import uk.govuk.app.design.ui.theme.GovUkTheme
-import uk.govuk.app.networking.R
+import uk.govuk.app.design.R
 
 @Composable
 fun OfflineMessage(
