+++ design/src/main/res/values/strings.xml
@@ -4,4 +4,13 @@
     <string name="gov_uk">GOV.UK</string>
     <string name="gov_uk_alt_text">GOV dot UK</string>
     <string name="opens_in_web_browser">Opens in web browser</string>
+
+    <!-- Error-->
+    <string name="no_internet_title">You are not connected to the internet</string>
+    <string name="no_internet_description">You need to have an internet connection to use the GOV.UK app.\n\nReconnect to the internet and try again.</string>
+    <string name="try_again">Try again</string>
+
+    <string name="problem_title">There’s a problem</string>
+    <string name="problem_description">This page is not working. Try again later, or go to the GOV.UK website.</string>
+    <string name="go_to_the_gov_uk_website">Go to the GOV.UK website</string>
 </resources>
