--- networking/src/main/kotlin/uk/govuk/app/networking/domain/Constants.kt
@@ -1,5 +1,0 @@
-package uk.govuk.app.networking.domain
-
-object Constants {
-    const val GOV_UK_URL = "https://www.gov.uk"
-}
